separate main file to 3 files
