--- policyengine_us/tests/policy/baseline/gov/local/co/denver/property_tax_relief/denver_property_tax_relief.yaml
@@ -1,83 +1,0 @@
-- name: Case 1, Homeowner with income higher than ami, ineligible. 
-  period: 2023
-  input:
-    spm_unit_size: 1 
-    spm_unit_net_income: 53_000
-    real_estate_taxes: 1_000
-    rent: 0
-    has_usda_elderly_disabled: true
-    tax_unit_dependents: 0 
-  output:
-    denver_property_tax_relief: 0 
-
-- name: Case 2, Homeowner with income less than ami, eligible. 
-  period: 2023
-  input:
-    spm_unit_size: 1 
-    spm_unit_net_income: 52_000
-    real_estate_taxes: 1_000
-    rent: 0
-    has_usda_elderly_disabled: true
-    tax_unit_dependents: 0 
-  output:
-    denver_property_tax_relief: 1_800
-
-- name: Case 3, renter with income higher than ami, ineligible. 
-  period: 2023
-  input:
-    spm_unit_size: 1 
-    spm_unit_net_income: 22_000
-    real_estate_taxes: 0
-    rent: 5_000
-    has_usda_elderly_disabled: true
-    tax_unit_dependents: 0 
-  output:
-    denver_property_tax_relief: 0  
-
-- name: Case 4, Renter with income less than ami, eligible. 
-  period: 2023
-  input:
-    spm_unit_size: 1 
-    spm_unit_net_income: 20_000
-    real_estate_taxes: 0
-    rent: 1_000
-    has_usda_elderly_disabled: true
-    tax_unit_dependents: 0 
-  output:
-    denver_property_tax_relief:  1_000
-
-- name: Case 5, renter household size of 3 with income less than ami, eligible. 
-  period: 2023
-  input:
-    spm_unit_size: 3 
-    spm_unit_net_income: 33_000
-    real_estate_taxes: 0
-    rent: 1_000
-    has_usda_elderly_disabled: true
-    tax_unit_dependents: 0 
-  output:
-    denver_property_tax_relief: 1_000 
-
-- name: Case 6, Homeowner with income less than ami, but is not elderly, not disabled, no dependent, hense ineligible.  
-  period: 2023
-  input:
-    spm_unit_size: 1 
-    spm_unit_net_income: 20_000
-    real_estate_taxes: 1_000
-    rent: 0
-    has_usda_elderly_disabled: false
-    tax_unit_dependents: 0 
-  output:
-    denver_property_tax_relief: 0 
-
-- name: Case 7, Renter with income less than ami, but is not elderly nor disabled, ineligible. 
-  period: 2023
-  input:
-    spm_unit_size: 1 
-    spm_unit_net_income: 20_000
-    real_estate_taxes: 0
-    rent: 1_000
-    has_usda_elderly_disabled: false
-    tax_unit_dependents: 1 
-  output:
-    denver_property_tax_relief:  0
