+++ policyengine_us/tests/policy/baseline/gov/local/co/denver/property_tax_relief/denver_property_tax_relief_renter.yaml
@@ -1,0 +1,39 @@
+- name: Case 1, renter with income higher than ami, ineligible. 
+  period: 2023
+  input:
+    spm_unit_size: 1 
+    spm_unit_net_income: 22_000
+    rent: 5_000
+    has_usda_elderly_disabled: true
+  output:
+    denver_property_tax_relief_renter: 0  
+
+- name: Case 2, Renter with income less than ami, eligible. 
+  period: 2023
+  input:
+    spm_unit_size: 1 
+    spm_unit_net_income: 20_000
+    rent: 1_000
+    has_usda_elderly_disabled: true
+  output:
+    denver_property_tax_relief_renter:  1_000
+
+- name: Case 3, renter household size of 3 with income less than ami, eligible. 
+  period: 2023
+  input:
+    spm_unit_size: 3 
+    spm_unit_net_income: 33_000
+    rent: 1_000
+    has_usda_elderly_disabled: true 
+  output:
+    denver_property_tax_relief_renter: 1_000 
+
+- name: Case 4, Renter with income less than ami, but is not elderly nor disabled, ineligible. 
+  period: 2023
+  input:
+    spm_unit_size: 1 
+    spm_unit_net_income: 20_0000
+    rent: 1_000
+    has_usda_elderly_disabled: false
+  output:
+    denver_property_tax_relief_renter:  0
